refactor(schemas): nest physics + skel under schemas/
