+++ src/schemas/skel/skeleton_query.rs
