+++ src/schemas/skel/skinning_query.rs
